feat: support and customization of retry for HTTP errors
- retrying was only based on error throwing, if an endpoint was successfully
answering, the request was not throwing an exception however the status might be retryable.
Tt is the case for Not Found, ENOTFOUND is retried, a successful 404 response is not.
This commit adds support and customization for such retry behaviour.
Replication delay tolerance (retrying 404) is an example use case for such granular customization of the "retry policy"

- this fix specifies no default retry function to be backward compatible

- for the sake of making the "retry mechanism" a bit more friendly, we exposed the max retries value